+++ ProductivityFeatures/Update163.cs
@@ -1,10 +1,13 @@
-﻿namespace ProductivityFeatures
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ProductivityFeatures
 {
     // Rename file when renaming a class, interface, or enum
     // Place your cursor in the class name and type (Ctrl + R, R) to 
     // open the Rename dialogue and check the ‘Rename file’ box
-    class Update163P1
+    class Update163
     {
-
     }
 }
